refactor: move ts auto-import caching code to `@volar/typescript-faster`
