+++ packages/typescript-faster/README.md
@@ -1,0 +1,3 @@
+# @volar/typescript-faster
+
+TypeScript LanguageService Completion API is slow, this package make it faster.
