+++ packages/typescript-faster/src/4_7/moduleSpecifierCache.ts
