+++ packages/typescript-faster/src/4_7/packageJsonCache.ts
