+++ packages/vue-typescript/src/typescriptRuntime.ts
@@ -2,7 +2,7 @@ import * as path from 'path';
 import type * as ts from 'typescript/lib/tsserverlibrary';
 import { LanguageServiceHost } from './types';
 import * as localTypes from './utils/localTypes';
-import { injectCacheLogicToLanguageServiceHost } from './utils/ts';
+import * as tsFaster from '@volar/typescript-faster';
 import { createVueFile, EmbeddedFile } from './vueFile';
 import { createVueFiles } from './vueFiles';
 
@@ -26,8 +26,8 @@ export function createTypeScriptRuntime(options: {
 	let lastProjectVersion: string | undefined;
 	let tsProjectVersion = 0;
 
-	if (!options.isVueTsc) {
-		injectCacheLogicToLanguageServiceHost(ts, tsLsHost, tsLsRaw);
+	if (!options.isVueTsc) { // not really needed disable this for vue-tsc
+		tsFaster.decorate(ts, tsLsHost, tsLsRaw);
 	}
 
 	return {
