+++ packages/vue-typescript/src/utils/ts.ts
@@ -1,92 +1,6 @@
 import type * as ts from 'typescript/lib/tsserverlibrary';
-import { createModuleSpecifierCache } from './moduleSpecifierCache';
-import { createPackageJsonCache, PackageJsonInfo, Ternary } from './packageJsonCache';
 import * as path from 'path';
 import type { VueCompilerOptions } from '../types';
-
-export function injectCacheLogicToLanguageServiceHost(
-	ts: typeof import('typescript/lib/tsserverlibrary'),
-	host: ts.LanguageServiceHost,
-	service: ts.LanguageService,
-) {
-
-	const versionNums = ts.version.split('.').map(s => Number(s));
-	const greaterThan47 = versionNums[0] > 4 || (versionNums[0] === 4 && versionNums[1] >= 7);
-	if (!greaterThan47) {
-		console.warn('TypeScript auto-import cache only working for TypeScript version < 4.7 on v0.35.0 later, please downgrade to v0.34.17 or lower for TypeScript version:', ts.version);
-		return;
-	}
-
-	const _createCacheableExportInfoMap = (ts as any).createCacheableExportInfoMap;
-	const _combinePaths = (ts as any).combinePaths;
-	const _forEachAncestorDirectory = (ts as any).forEachAncestorDirectory;
-	const _getDirectoryPath = (ts as any).getDirectoryPath;
-	const _toPath = (ts as any).toPath;
-	const _createGetCanonicalFileName = (ts as any).createGetCanonicalFileName;
-
-	if (
-		!_createCacheableExportInfoMap
-		|| !_combinePaths
-		|| !_forEachAncestorDirectory
-		|| !_getDirectoryPath
-		|| !_toPath
-		|| !_createGetCanonicalFileName
-	) return;
-
-	const moduleSpecifierCache = createModuleSpecifierCache();
-	const exportMapCache = _createCacheableExportInfoMap({
-		getCurrentProgram() {
-			return service.getProgram();
-		},
-		getPackageJsonAutoImportProvider() {
-			return service.getProgram();
-		},
-		getGlobalTypingsCacheLocation() {
-			return undefined;
-		},
-	});
-	const packageJsonCache = createPackageJsonCache(ts, {
-		...host,
-		// @ts-expect-error
-		host: { ...host },
-		toPath,
-	});
-
-	// @ts-expect-error
-	host.getCachedExportInfoMap = () => exportMapCache;
-	// @ts-expect-error
-	host.getModuleSpecifierCache = () => moduleSpecifierCache;
-	// @ts-expect-error
-	host.getPackageJsonsVisibleToFile = (fileName: string, rootDir?: string) => {
-		const rootPath = rootDir && toPath(rootDir);
-		const filePath = toPath(fileName);
-		const result: PackageJsonInfo[] = [];
-		const processDirectory = (directory: ts.Path): boolean | undefined => {
-			switch (packageJsonCache.directoryHasPackageJson(directory)) {
-				// Sync and check same directory again
-				case Ternary.Maybe:
-					packageJsonCache.searchDirectoryAndAncestors(directory);
-					return processDirectory(directory);
-				// Check package.json
-				case Ternary.True:
-					const packageJsonFileName = _combinePaths(directory, "package.json");
-					// this.watchPackageJsonFile(packageJsonFileName as ts.Path); // TODO
-					const info = packageJsonCache.getInDirectory(directory);
-					if (info) result.push(info);
-			}
-			if (rootPath && rootPath === directory) {
-				return true;
-			}
-		};
-
-		_forEachAncestorDirectory(_getDirectoryPath(filePath), processDirectory);
-		return result;
-	};
-
-	function toPath(fileName: string) {
-		return _toPath(fileName, host.getCurrentDirectory(), _createGetCanonicalFileName(host.useCaseSensitiveFileNames?.()));
-	}
-}
 
 export function createParsedCommandLine(
 	ts: typeof import('typescript/lib/tsserverlibrary'),
